Big re-work how the docker container builds prod vs dev * Don't build prod on boot, build it during docker build * Remove copied assets / keycloak assets after build is done in container * Volume mount keycloak to assets/keycloak for dev * Don't volume link /static
This should make the container much faster to boot, as it will no longer
need to build every time we boot it. It will also clear up junk data
left over from the build in the container.

This shouldn't affect much, and I've tested dev and prod builds of the
container but we'll see how the CI fairs.